requirements.txt
appdirs==1.4.4
arrow==0.16.0
asgiref==3.2.10
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
distlib==0.3.1
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
filelock==3.0.12
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.2.0
pipenv==2020.8.13
poyo==0.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
virtualenv==20.0.33
virtualenv-clone==0.5.4
django-cors-headers==3.5.0
